Organizations (#132)

+++ node/src/gql/mutation/me/import-task-delete.ts
@@ -14,8 +14,8 @@ import {
   Mentor as MentorModel,
   ImportTask as ImportTaskModel,
 } from '../../../models';
-import { Mentor } from '../../../models/Mentor';
-import { User, UserRole } from '../../../models/User';
+import { User } from '../../../models/User';
+import { canEditMentor } from '../../../utils/check-permissions';
 
 export const importTaskDelete = {
   type: GraphQLBoolean,
@@ -29,20 +29,12 @@ export const importTaskDelete = {
     },
     context: { user: User }
   ): Promise<boolean> => {
-    let mentor: Mentor = await MentorModel.findOne({
-      user: context.user._id,
-    });
+    const mentor = await MentorModel.findById(args.mentorId);
     if (!mentor) {
-      throw new Error('you do not have a mentor');
+      throw new Error('invalid mentor');
     }
-    if (args.mentorId && `${mentor._id}` !== `${args.mentorId}`) {
-      if (
-        context.user.userRole !== UserRole.ADMIN &&
-        context.user.userRole !== UserRole.CONTENT_MANAGER
-      ) {
-        throw new Error('you do not have permission to edit this mentor');
-      }
-      mentor = await MentorModel.findById(args.mentorId);
+    if (!(await canEditMentor(mentor, context.user))) {
+      throw new Error('you do not have permission to edit this mentor');
     }
     const taskDelete = await ImportTaskModel.deleteOne({
       mentor: mentor._id,
